resume-parser/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx
spacy